--- old-remark-cite/package.json
@@ -1,61 +1,0 @@
-{
-	"name": "@benrbray/remark-cite",
-	"version": "1.1.0",
-	"license": "MIT",
-	"description": "A `remark` plugin for pandoc-style citations.",
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/benrbray/remark-cite.git",
-		"directory": "remark-cite"
-	},
-	"keywords": [
-		"markdown",
-		"pandoc",
-		"citations",
-		"bibliography",
-		"remark",
-		"remark-plugin",
-		"micromark",
-		"micromark-plugin",
-		"micromark-extension"
-	],
-	"author": {
-		"name": "Benjamin R. Bray",
-		"url": "https://benrbray.com/"
-	},
-	"main": "dist/index.cjs.js",
-	"module": "dist/index.esm.js",
-	"browser": "dist/index.umd.js",
-	"files": [
-		"dist"
-	],
-	"scripts": {
-		"build": "rollup -c",
-		"clean": "npx del dist",
-		"prepare": "npm run clean; npm run build",
-		"pretest": "npm run clean; npm run build",
-		"test": "env TS_NODE_COMPILER_OPTIONS='{\"module\": \"commonjs\" }' mocha -r ts-node/register 'test/**/*.ts'",
-		"dev:fix-links": "npm link @benrbray/micromark-extension-cite @benrbray/mdast-util-cite"
-	},
-	"devDependencies": {
-		"@babel/plugin-transform-runtime": "^7.13.15",
-		"@rollup/plugin-babel": "^5.3.0",
-		"@rollup/plugin-commonjs": "^18.1.0",
-		"@types/mdast": "^3.0.3",
-		"@types/mocha": "^8.2.2",
-		"@types/unist": "^2.0.3",
-		"@wessberg/rollup-plugin-ts": "^1.3.14",
-		"del-cli": "^3.0.1",
-		"mdast-util-from-markdown": "^0.8.5",
-		"mocha": "^8.3.2",
-		"remark-parse": "^9.0.0",
-		"remark-stringify": "^9.0.1",
-		"rollup": "^2.47.0",
-		"ts-node": "^9.1.1",
-		"unified": "^9.2.1"
-	},
-	"dependencies": {
-		"@benrbray/mdast-util-cite": "^1.0.1",
-		"@benrbray/micromark-extension-cite": "^1.0.0"
-	}
-}
