+++ remark-cite/package.json
@@ -1,16 +1,61 @@
 {
-  "name": "remark-cite",
-  "private": true,
-  "version": "0.0.0",
+  "name": "@benrbray/remark-cite",
+  "version": "1.1.0",
+  "license": "MIT",
+  "description": "A `remark` plugin for pandoc-style citations.",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/benrbray/remark-cite.git",
+    "directory": "remark-cite"
+  },
+  "keywords": [
+    "markdown",
+    "pandoc",
+    "citations",
+    "bibliography",
+    "remark",
+    "remark-plugin"
+  ],
+  "author": {
+    "name": "Benjamin R. Bray",
+    "url": "https://benrbray.com/"
+  },
   "type": "module",
+  "exports": {
+    ".": "./dist/remark-cite.js"
+  },
+  "types": "dist/remark-cite.d.ts",
+  "files": [
+    "dist",
+    "lib"
+  ],
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
-    "preview": "vite preview"
+    "clean": "pnpm del dist",
+    "prepare": "pnpm run clean; pnpm run build",
+    "pretest": "pnpm run clean; pnpm run build",
+    "preview": "vite preview",
+    "test": "vitest"
+  },
+  "dependencies": {
+    "@benrbray/mdast-util-cite": "workspace:^1.0.1",
+    "@benrbray/micromark-extension-cite": "workspace:^1.0.0"
   },
   "devDependencies": {
+    "@types/mdast": "^4.0.4",
+    "@types/unist": "^3.0.2",
+    "del-cli": "^5.1.0",
+    "mdast-util-from-markdown": "^2.0.1",
+    "mdast-util-to-markdown": "^2.1.0",
+    "micromark-util-types": "^2.0.0",
+    "remark-parse": "^11.0.0",
+    "remark-stringify": "^11.0.0",
     "typescript": "^5.2.2",
+    "unified": "^11.0.4",
     "vite": "^5.2.0",
-    "vite-tsconfig-paths": "^4.3.2"
+    "vite-plugin-dts": "^3.9.1",
+    "vite-tsconfig-paths": "^4.3.2",
+    "vitest": "^1.6.0"
   }
 }
