+++ remark-cite/src/main.ts
@@ -1,9 +1,7 @@
 import './style.css'
 
-import { number } from "@lib/main";
-
 document.querySelector<HTMLDivElement>('#app')!.innerHTML = `
   <div>
-    Hello! ${number}
+    Hello!
   </div>
 `
